% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_split_or_merge_event.R
\name{analyze_split_or_merge_event}
\alias{analyze_split_or_merge_event}
\title{Analyze split or merge event}
\usage{
analyze_split_or_merge_event(
  i,
  events,
  xs,
  ys,
  timestamps,
  max_time = 600,
  thresh_h = 50,
  thresh_l = 15,
  time_window = 300,
  plot = T,
  seconds_per_time_step = 1
)
}
\arguments{
\item{i}{index of the row to use in the events table}

\item{events}{a data frame of fission-fusion events, output from \code{identify_splits_and_merges}}

\item{xs}{\verb{n_inds x n_times} matrix of x coordinates}

\item{ys}{\verb{n_inds x n_times} matrix of y coordinates}

\item{timestamps}{vector of timestamps of length \code{n_times} in datetime format}

\item{max_time}{maximum time steps forward and back to look for the start and end of the event (units are timesteps)}

\item{thresh_h}{upper threshold for determining when subgroups are "apart" (default 50)}

\item{thresh_l}{lower threshold for determining when subgroups are "together" (default 15)}

\item{time_window}{time steps to move backward or forward in time to identify the before and after times}

\item{plot}{whether to plot the event (if \code{T}) or not (if \code{F})}

\item{seconds_per_time_step}{seconds per time step (default 1)}
}
\value{
Returns a list (\code{out}) of information extracted about the event, as well as a plot (if \code{plot = T}).

The list contains:

\code{start_time}: start time index

\code{end_time}: end time index

\code{before_time}: before time index

\code{after_time}: after time index

\code{disps}: matrix of displacements of the different subgroups (rows) during the different
time intervals (columns). Rows and columns are named for easy access.

\code{speeds}: same format as \code{disps} matrix, but with speeds, in m / s
\code{split_angle}: split angle in degrees, description above
\code{turn_angle_A}: turning angle for subgroup A in degrees, description above
\code{turn_angle_B}: turn angle for subgroup B in degrees, description above
The plot shows (top) dyadic distance over time with lines showing the identified times
and (bottom) a visualization of trajectories of the two subgroups,
}
\description{
Analyze (and, if \code{plot=T} make a visualization of) a fission-fusion event.
}
\details{
This function takes in information about a fission or fusion event as well as
tracking data to identify relevant time points in the fission or fusion event and compute relevant metrics about the event.
First, the \code{start_time} and \code{end_time} of the event are determined based on
a double threshold method, then a \code{before_time} and \code{after_time} are identified.
This identifies 3 phases of the event: before (\code{before_time:start_time}), during
(\code{start_time:end_time}) and after (\code{end_time:after_time}).
Finally the displacements and speeds of the subgroups during these phases and various
relevant angles are calculated. More details are given below. Note that this function can only
consider events involving 2 subgroups.
}
\section{Additional details}{


\emph{How are the start and end time identified?}

As a start, we look at a window of time around an identified event (can be
manually or automatically identified). The size of the window is determined by
the parameter '\code{max_time}', and we go from \code{(tidx - max_time):(tidx + max_time)}
where \code{tidx} is the identified event time index.

Within the window, we compute the distance between the centroids of the two
subgroups that are splitting or merging at each time point.
We then use a double threshold method to determine the start and end point of
the fission or fusion event within that time window. To do so, we first
categorize all time points as being above the higher threshold \code{thresh_h} (2),
between the two thresholds (1), or below the lower threshold \code{thresh_l} (0). We
then identify contiguous periods of time where the dyadic distance went from
2-111111(any number of ones)-0 (i.e. high-mid-low) for a fusion or 0-1111...-2
(i.e. low-mid-high) for a fission. If there are multiple possible time periods
detected within the window, we choose the one that is closest to \code{tidx}. (See also
subtlety 1 below).

\emph{How are the before and after times identified?}

The before time is defined as the time point \code{tidx - time_window} (default
\code{time_window = 300} sec), unless the two groups are not sufficiently together (for
a fission) or apart (for a fusion) at that time. If the latter, the \code{before_time}
is identified as the point just before the two subgroups cross a threshold midway
between \code{thresh_l} and \code{thresh_h} (i.e. at \code{(thresh_l + thresh_h) / 2}). The logic here
is that, for a fusion we are looking for what the full group (combination of the
two eventual subgroups) was doing before they began to split. However, we do not
want this point to fall during a prior fusion event, so we require the centroids
of the two subgroups to be less than \code{(thresh_l + thresh_h) / 2} distance apart.
The after time is defined in an analogous way, but using the time period after the
event. It is usually set to \code{tidx + time_window}, unless the subgroups come back too
close together (for a fission) or go too far apart (for a fusion). Again, we use
the time just prior to crossing the midway point between the two thresholds as
the \code{after_time}, if this threshold is crossed before \code{time_window} seconds has elapsed.

\emph{How are the displacements defined?}

We compute the displacement of the centroid of each subgroup (group A and group B)
as well as the displacement of the centroid of the combined group (group AB) during
each of the time windows: before (\code{before_time:start_time}), during (\code{start_time:end_time})
and after (\code{end_time:after_time}).

\emph{How are the angles defined?}

We define 3 relevant angles relevant to a fission event (might define more later
for merge events):
\code{split_angle}: this is the angle at which the two groups split. It is defined as
the angle traced out by the points \code{p_A(end_time)}, \code{p_AB(start_time)}, and \code{p_B(end_time)}
where \code{p_A(t)} is the position of subgroup A's centroid at time \code{t} and likewise
for subgroup B and the combined group.
\code{turn_angle_A}: the turning angle of subgroup A. This is defined as the angle
formed by the points \code{p_AB(before_time)}, \code{p_AB(start_time)}, and \code{p_A(end_time)}
turn_angle_B: likewise for subgroup B
Note that all angles use the point \code{p_AB(start_time)} as their central point

\emph{SUBTLETIES}:

\emph{Subtlety 1}: Sometimes, due to the multi-scale nature of these events and the
fact that we are approximating subgroup locations with centroids, the dyadic
distance does not go below the lower threshold \code{thresh_l} and/or above the upper
threshold \code{thresh_h}. In this case, we still try to identify the \code{start_time} and
end_time, but modify the thresholds as follows:

First, let's define the period between \code{tidx - max_time} and \code{tidx} as the \emph{prior period},
the period between \code{tidx}  and \code{tidx + max_time} as the \emph{subsequent period}, and the
period between \code{(tidx - max_time/2)} and \code{(tidx + max_time/2)} as the \emph{middle period}.

For a fission, if the dyadic distance does not go above \code{thresh_h} in the \emph{subsequent period},
then we instead replace \code{thresh_h} with the maximum - .001 of the dyadic
distance during that time period. Second, if the dyadic distance during the \emph{middle period}
does not drop below \code{thresh_l}, then we instead move \code{thresh_l} to the minimum + .001
of the dyadic distance during the \emph{middle period}.

For a fusion, everything is reversed. If the dyadic distance does not go
above \code{thresh_h} during the \emph{prior period}, we replace \code{thresh_h} with the maximum - .001
of the dyadic distance during the \emph{subsequent period}. And if the dyadic distance does
not go below \code{thresh_l} during the \emph{middle period}, we replace \code{thresh_l} with the minimum + .001
of the dyadic distance during the \emph{middle period}.

In very rare cases, due to these rules the upper bound may get changed to
something below the original \code{thresh_l}, or the lower bound may get changed to
something above the original \code{thresh_h}. In that case, we revert to the original thresholds.

\emph{Subtlety 2}: NA handling. NAs are handled in a couple of different ways:

For finding the \code{start_time} and \code{end_time}, NAs are essentially ignored, and cannot
be part of the sequence from \code{start_time:end_time} (at least not in terms of the)
dyadic distance... individuals can drop out and they will just be excluded from
the centroid calculation.
If no start and end times are found, nothing else is computed (all output values are
filled in with \code{NA}s or \code{NULL} for matrices).

For finding the \code{before_time} and \code{after_time}, if an \code{NA} is hit in the forward or
backward direction, the \code{before_time} (or respectively, the \code{after_time}) is marked
as \code{NA}. Metrics involving that time point can then not be computed and are also set to
\code{NA}.

If the \code{start_time} and \code{end_time} are on different days, then both are given \code{NA}. TODO: look into fixing this if have continuous data

If the \code{before_time} and \code{start_time}, or \code{after_time} and \code{end_time}, are on different days, then
both are given NA.

If the \code{before_time} or \code{after_time} are \code{NA}, then other metrics stemming from those times get \code{NA}
}

\author{
Ariana Strandburg-Peshkin (primary author)

NOT YET CODE REVIEWED
}
