% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocomo-package.R
\docType{package}
\name{cocomo-package}
\alias{cocomo}
\alias{cocomo-package}
\title{cocomo: Communication and Collective Movement Package}
\description{
This is a package developed in the CoCoMo group and the Department for the Ecology of Animal Societies at the Max Planck Institute of Animal Behavior to analyze communication and movement data of animal groups.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://livingingroups.github.io/cocomo/}
}

}
\keyword{internal}
