% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_movement_and_calls_visualization.R
\name{generate_movement_and_calls_visualization}
\alias{generate_movement_and_calls_visualization}
\title{Make a visualization of individual positions and calls during a time period specified by the user.
The visualization can either show all individuals in the provided data or can zoom in on one individual
and show its behavior in detail.}
\usage{
generate_movement_and_calls_visualization(
  xs = NULL,
  ys = NULL,
  timestamps = NULL,
  calls = NULL,
  start_time = NULL,
  end_time = NULL,
  time_step = 1,
  output_dir = NULL,
  ind_names = NULL,
  focus_ind = NULL,
  zoom_radius = 400,
  bg_color = "black",
  satellite_map = TRUE,
  utm_epsg = NULL,
  auto_zoom_tiles = TRUE,
  zoom_global = 15,
  zoom_focus = NULL,
  colors_inds = NULL,
  colors_calls = NULL,
  ind_point_size = 2,
  call_point_size = 1,
  pchs_inds = NULL,
  pchs_calls = NULL,
  tail_time = 10,
  call_persist_time = 10,
  show_legend_inds = TRUE,
  sort_legend_inds = NULL,
  show_legend_calls = FALSE,
  legend_loc = "topright",
  scalebar_size = 100,
  events = NULL,
  highlighted_radius = 1000
)
}
\arguments{
\item{xs}{matrix of dimensions \code{n_inds} x \code{n_times} where \code{xs[i,t]} gives the x position (numeric) of individual \code{i} at time step \code{t}, positions in UTM format}

\item{ys}{matrix of dimensions \code{n_inds} x \code{n_times} where \code{ys[i,t]} gives the y position (numeric) of individual \code{i} at time step \code{t}, positions in UTM format}

\item{timestamps}{vector of timestamps (POSIXct), must have same dimensions as columns of \code{xs} and \code{ys} matrices}

\item{calls}{data frame where first column (\code{'ind_idx'}) specifies the index of the individual that gave the call, second column (\code{'time_idx'}) specifies the time index at which the call was given, and third column (\code{'call_type'}) specifies the type of call (character string)}

\item{start_time}{time index at which to start the video}

\item{end_time}{time index at which to end the video}

\item{time_step}{time step to use (an image every time_step timesteps will be produced)}

\item{output_dir}{directory in which to store the folder of outputted images}

\item{ind_names}{vector of names of the individuals}

\item{focus_ind}{name or index of a focal individual. This centers the frame on one individual instead of producing a broad overview. When using name, ind_names must be provided as well.}

\item{zoom_radius}{meters around the focal individual (UTM)}

\item{bg_color}{background color of the plot ("black", "white", etc.)}

\item{satellite_map}{use a satellite map as the background (T or F)}

\item{utm_epsg}{WGS 84 EPSG reference number to find the geographic location of the satellite map tile, use https://epsg.io to find the reference number}

\item{auto_zoom_tiles}{choose satellite map zoom automatically when focusing on individual, determines satellite map resolution}

\item{zoom_global}{zoom level for global tiles of the background satellite map, determines satellite map resolution}

\item{zoom_focus}{optional explicit map zoom for focus tiles, determines satellite map resolution}

\item{colors_inds}{vector of colors to use for each individual (length \code{n_inds}). Color indices must match the individual index in calls , xs and ys}

\item{colors_calls}{vector of colors to use for each call type (alphabetical order by call_type)}

\item{ind_point_size}{size of the individual points}

\item{call_point_size}{size of the points for calls}

\item{pchs_inds}{vector of plotting symbols for individuals}

\item{pchs_calls}{vector of plotting symbols for calls}

\item{tail_time}{number of previous time steps to plot as a "tail" which trails the point showing the current location}

\item{call_persist_time}{number of previous time steps to still show the calls (they will shrink linearly over time in size)}

\item{show_legend_inds}{whether to plot a legend showing the names of the individuals (T or F)}

\item{sort_legend_inds}{vector to sort the legend showing the names of the individuals (vector containing the sort order by index)}

\item{show_legend_calls}{whether to plot a legend showing the names of the call types (T or F)}

\item{legend_loc}{location of the individual legend, either 'topleft' or 'topright'. Call legend will be opposite}

\item{scalebar_size}{number of meters for the scalebar, set to 0 to deactivate}

\item{events}{data frame with columns \code{event_id}, \code{start_time_idx},\code{end_time_idx},\code{initiator}}

\item{highlighted_radius}{radius of the highlighted location (usually an epicenter from hyena whoop analysis)}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\author{
Ariana Strandburg-Peshkin (primary author)

Marius Faiß (secondary author)

NOT YET CODE REVIEWED
}
