% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_angle_between_vectors.R
\name{get_angle_between_vectors}
\alias{get_angle_between_vectors}
\title{Get angle between vectors in radians or degrees}
\usage{
get_angle_between_vectors(
  x1_i,
  y1_i,
  x1_f,
  y1_f,
  x2_i,
  y2_i,
  x2_f,
  y2_f,
  degrees = F
)
}
\arguments{
\item{x1_i}{x coordinate of initial point of vector 1}

\item{y1_i}{y coordinate of initial point of vector 1}

\item{x1_f}{x coordinate of final point of vector 1}

\item{y1_f}{y coordinate of final point of vector 1}

\item{x2_i}{x coordinate of initial point of vector 2}

\item{y2_i}{y coordinate of initial point of vector 2}

\item{x2_f}{x coordinate of final point of vector 2}

\item{y2_f}{y coordinate of final point of vector 2}

\item{degrees}{if \code{T}, angle will be returned in degrees (otherwise it will be returned in radians)}
}
\value{
Returns the angle between the two specified vectors
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Get the angle between two vectors specified by their end points using law of cosines.
Vector 1 is defined as the vector pointing from the point \verb{(x1_i, y1_i)} to \verb{(x1_f, y1_f)} and
vector 2 is defined as the vector point from the point \verb{(x2_i, y2_i)} to \verb{(x2_f, y2_f)}.
The angle is defined as the angle produced if the two vectors are joined at the initial
endpoints (rather than head-to-tail). The resulting angle is always positive.
}
\author{
Ariana Strandburg-Peshkin

NOT YET CODE REVIEWED
}
