% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_group_centroid.R
\name{get_group_centroid}
\alias{get_group_centroid}
\title{Get group centroid}
\usage{
get_group_centroid(xs, ys, min_inds_tracked = NULL)
}
\arguments{
\item{xs}{\verb{N x n_times} matrix giving x coordinates of each individual over time}

\item{ys}{\verb{N x n_times} matrix giving y coordinates of each individual over time}

\item{min_inds_tracked}{if specified, sets a minimum number of individuals that must be tracked at any moment in time to compute centroid (otherwise the centroid will be NA at that time point}
}
\value{
Returns coordinates of the group centroid over time, an \verb{n_times x 2} matrix of x (col 1) and y (col 2) coordinates
}
\description{
Get the group's centroid (mean x and y position) over time.
If specified, only compute if at least \code{min_inds_tracked} are tracked in a given
time step.
}
\author{
Ariana Strandburg-Peshkin (primary author)

NOT YET CODE REVIEWED
}
