% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_group_dyadic_distances.R
\name{get_group_dyadic_distances}
\alias{get_group_dyadic_distances}
\title{Get group dyadic distances}
\usage{
get_group_dyadic_distances(xs, ys)
}
\arguments{
\item{xs}{\verb{N x n_times} matrix giving x coordinates of each individual over time}

\item{ys}{\verb{N x n_times} matrix giving y coordinates of each individual over time}
}
\value{
\verb{N x N x n_times} \code{dyad_dists} array giving dyadic distance between all pairs of individuals at each time step
\code{dyad_dists[i,j,t]} gives the distance between individuals \code{i} and \code{j} at time \code{t}
}
\description{
Computes the distance between each pair of individuals over time.
}
\author{
Ariana Strandburg-Peshkin (primary author)

NOT YET CODE REVIEWED
}
