% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_group_elongation.R
\name{get_group_elongation}
\alias{get_group_elongation}
\title{Get group polarization}
\usage{
get_group_elongation(xs, ys, min_inds_tracked = 3)
}
\arguments{
\item{xs}{\verb{N x n_times} matrix giving x coordinates of each individual over time}

\item{ys}{\verb{N x n_times} matrix giving y coordinates of each individual over time}

\item{min_inds_tracked}{if specified, sets a minimum number of individuals that must be tracked at any moment in time to compute elongation (otherwise the elongation will be NA at that time point). Default is 3 (below this, elongation is generally undefined)}
}
\value{
Returns a list containing:
\code{out$elongation}: a vector of length \code{n_times} giving the elongation of the group over time
\code{out$long_axis_angle}: a vector of length \code{n_times} giving the angle of the long axis over time
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Computes the elongation of the group at each time step \code{t} as well as the long axis vector.
The elongation is a measure of how "line-link" vs "circle-like" the group is, ranging from 0 (completely circular) to 1 (completely elongated)
The elongation is defined by first performing a PCA on the x/y data of the group. The axis of greatest variation is defined
as the 'long axis' and the axis perpendicular to this is defined as the 'short axis'. The elonation is then
defined as 1 - (length of short axis / length of long axis).
}
\author{
Ariana Strandburg-Peshkin (primary author)

NOT YET CODE REVIEWED
}
