% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_group_heading_and_speed.R
\name{get_group_heading_and_speed}
\alias{get_group_heading_and_speed}
\title{Get group heading}
\usage{
get_group_heading_and_speed(
  xs,
  ys,
  heading_type,
  spatial_R = NULL,
  t_window = NULL,
  forward = T,
  min_inds_tracked = NUL,
  seconds_per_time_step = 1
)
}
\arguments{
\item{xs}{\verb{N x n_times} matrix giving x coordinates of each individual over time}

\item{ys}{\verb{N x n_times} matrix giving y coordinates of each individual over time}

\item{heading_type}{character string specifying heading type - \code{'spatial'} or \code{'temporal'}}

\item{spatial_R}{radius to use for spatial headings (if \code{heading_type = 'spatial'})}

\item{t_window}{temporal window to use for temporal headings (if \code{heading_type = 'temporal'})}

\item{forward}{whether to compute headings into the future (\code{forward = T}) or the past (\code{forward = F})}

\item{min_inds_tracked}{if specified, sets a minimum number of individuals that must be tracked at any moment in time to compute heading (otherwise the heading will be NA at that time point}

\item{seconds_per_time_step}{number of seconds corresponding to each time step}
}
\value{
Returns the group heading over time, a vector of length \code{n_times}
}
\description{
Computes the group centroid heading over time using either temporal or spatial headings, and
going either forward (into the future) or backward (into the past) when computing the heading
}
\author{
Ariana Strandburg-Peshkin (primary author)

NOT YET CODE REVIEWED
}
