% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_group_polarization.R
\name{get_group_polarization}
\alias{get_group_polarization}
\title{Get group polarization}
\usage{
get_group_polarization(xs, ys)
}
\arguments{
\item{xs}{\verb{N x n_times} matrix giving x coordinates of each individual over time}

\item{ys}{\verb{N x n_times} matrix giving y coordinates of each individual over time}
}
\value{
Returns a vector of length \code{n_times} giving the polarization of the group over time
(where the heading is undefined, returns NA)
}
\description{
Computes the polarization of the group at each time step \code{t}.
The polarization is a measure of how aligned the group is, ranging from 0 (completely unaligned) to 1 (completely aligned)
The polarization is defined by adding up (vector addition) all of the heading vectors of all individuals at a given moment in time,
taking the length of the resultant vector, and dividing this by the number of individuals.
}
\author{
Ariana Strandburg-Peshkin (primary author)

NOT YET CODE REVIEWED
}
