% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_heading_and_speed_spatial.R
\name{get_heading_and_speed_spatial}
\alias{get_heading_and_speed_spatial}
\title{Get spatially discretized heading and speed over time}
\usage{
get_heading_and_speed_spatial(
  x_i,
  y_i,
  R,
  t_idxs = 1:length(x_i),
  forward = T,
  seconds_per_time_step = 1
)
}
\arguments{
\item{x_i}{vector of x coordinates for the trajectory}

\item{y_i}{vector of y coordinates for the trajectory}

\item{R}{radius used to compute the headings}

\item{t_idxs}{time indexes at which to compute the headings, speed, dts (defaults to entire trajectory)}

\item{forward}{whether to go forward in time from current position (if T), or backward (if F) when computing headings}

\item{seconds_per_time_step}{number of seconds corresponding to each time step}
}
\value{
Returns a list containing
\itemize{
\item \verb{$heads}: a time series of the heading of the individual (a vector of the same length as x_i and y_i), in radians, computed based on
spatial discretization around the time point
\item \verb{$speeds}: a time series of the average speed of the individual from the first point outside radius R to the current point
\item \verb{$dts}: time differences between the current point and the first point outside of the radius R (either forward or backward in time)
}

vector of spatially discretized headings, computed at all times or at times t_idxs if specified (other times are then filled in with NAs)
}
\description{
Gets spatially discretized heading for a given individual trajectory
This is defined as the unit vector pointing from the individual's current location at time \code{t} to
its location after it mas moved a distance of at least \code{R}.
}
\author{
Ariana Strandburg-Peshkin (primary author)

Reviewed by Brock
}
