% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_heading_and_speed_temporal.R
\name{get_heading_and_speed_temporal}
\alias{get_heading_and_speed_temporal}
\title{Get heading and speed over time (temporal)}
\usage{
get_heading_and_speed_temporal(
  x_i,
  y_i,
  t_window = 1,
  forward = T,
  seconds_per_time_step = 1
)
}
\arguments{
\item{x_i}{x coordinates of the individual (a vector whose length is the number of timesteps) or of a group centroid}

\item{y_i}{y coordinates of the individual (a vector whose length is the number of timesteps) or of the group centroid}

\item{t_window}{the time window over which to compute speed (number of time steps into the future or past)}

\item{forward}{a boolean value (defaults to T) indicating whether to compute speeds forward in time (if T) or backward in time (if F)}

\item{seconds_per_time_step}{number of seconds corresponding to each time step}
}
\value{
Returns a list containing \verb{$heads}: a time series of the heading of the individual (a vector of the same length as x_i and y_i), in radians,
\verb{$speeds}: a time series of the speed of the individual at each time point, and
\verb{$dts}: time differences between each point (will be all the same value, but given as a vector for compatibility with spatial headings function)
}
\description{
Gets the headings and speeds of an individual over time given a trajectory.
Uses a temporal window \code{t_window} (in units of timesteps) to compute the headings and speeds,
and can compute them based on future movement (\code{forward = T}) or past movement (\code{forward = F}).
Headings are defined as \code{x_head = dx / (dx^2 + dy^2)} and \code{y_head = dy / (dx^2 + dy^2)}
where, \code{dx[t] = x_i[t + t_window] - x_i[t]} and \code{dy[t] = y_i[t + t_window] - y_i[t]} for \code{forward = T}
and \code{dx[t] = x_i[t] - x_i[t - t_window]} and \code{dy[t] = y_i[t] - y_i[t - t_window]} for \code{forward = F}
Speeds are defined as \code{speed = sqrt(dx^2 + dy^2) / t_window}.
}
\author{
Ariana Strandburg-Peshkin

NOT YET CODE REVIEWED
}
