% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_headings_spatial.R
\name{get_headings_spatial}
\alias{get_headings_spatial}
\title{Get spatially discretized headings}
\usage{
get_headings_spatial(x_i, y_i, R, t_idxs = 1:length(x_i), forward = T)
}
\arguments{
\item{x_i}{vector of x coordinates for the trajectory}

\item{y_i}{vector of y coordinates for the trajectory}

\item{R}{radius used to compute the headings}

\item{forward}{whether to go forward in time from current position (if T), or backward (if F) when computing headings}

\item{t_idxs:}{time indexes at which to compute the headings (defaults to entire trajectory)}
}
\value{
vector of spatially discretized headings, computed at all times or at times t_idxs if specified (other times are then filled in with NAs)
}
\description{
Gets spatially discretized heading for a given individual trajectory
This is defined as the unit vector pointing from the individual's current location at time \code{t} to
its location after it mas moved a distance of at least \code{R}.
}
\author{
Ariana Strandburg-Peshkin (primary author)

NOT YET CODE REVIEWED
}
