% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_headings_temporal.R
\name{get_headings_temporal}
\alias{get_headings_temporal}
\title{Get headings (temporal)
Gets the headings of an individual over time given a trajectory.
Uses a temporal window \code{t_window} (in units of timesteps) to compute the headings, and can compute them
based on future movement (\code{forward = T}) or past movement (\code{forward = F}).
Headings are defined as \code{x_head = dx / (dx^2 + dy^2)} and \code{y_head = dy / (dx^2 + dy^2)}
where, \code{dx[t] = x_i[t + t_window] - x_i[t]} and \code{dy[t] = y_i[t + t_window] - y_i[t]} for \code{forward = T}
and \code{dx[t] = x_i[t] - x_i[t - t_window]} and \code{dy[t] = y_i[t] - y_i[t - t_window]} for \code{forward = F}}
\usage{
get_headings_temporal(x_i, y_i, t_window = 1, forward = T)
}
\arguments{
\item{x_i}{x coordinates of the individual (a vector whose length is the number of timesteps) or of a group centroid}

\item{y_i}{y coordinates of the individual (a vector whose length is the number of timesteps) or of the group centroid}

\item{t_window}{the time window over which to compute speed (number of time steps into the future or past)}

\item{forward}{a boolean value (defaults to T) indicating whether to compute speeds forward in time (if T) or backward in time (if F)}
}
\value{
Returns \code{heads}: a time series of the heading of the individual (a vector of the same length as x_i and y_i), in radians
}
\description{
Get headings (temporal)
Gets the headings of an individual over time given a trajectory.
Uses a temporal window \code{t_window} (in units of timesteps) to compute the headings, and can compute them
based on future movement (\code{forward = T}) or past movement (\code{forward = F}).
Headings are defined as \code{x_head = dx / (dx^2 + dy^2)} and \code{y_head = dy / (dx^2 + dy^2)}
where, \code{dx[t] = x_i[t + t_window] - x_i[t]} and \code{dy[t] = y_i[t + t_window] - y_i[t]} for \code{forward = T}
and \code{dx[t] = x_i[t] - x_i[t - t_window]} and \code{dy[t] = y_i[t] - y_i[t - t_window]} for \code{forward = F}
}
\author{
Ariana Strandburg-Peshkin

NOT YET CODE REVIEWED
}
