% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_local_density.R
\name{get_local_density}
\alias{get_local_density}
\title{Get local density}
\usage{
get_local_density(xs, ys, i, R)
}
\arguments{
\item{xs}{\verb{N x n_times} matrix giving x positions of all \code{N} individuals over \code{n_times} timesteps}

\item{ys}{\verb{N x n.times} matrix giving y positions of all \code{N} individuals over \code{n_times} timesteps}

\item{i}{the focal individual (for whom the local density will be computed)}

\item{R}{the radius over which to compute the local density (in same units as \code{xs} and \code{ys})}
}
\value{
Vector of length \code{n_times} giving the local density around the focal individual (number of individuals within a radius R) at each time point
}
\description{
Gets the number of individuals within a radius \code{R} of a focal individual \code{i} at each time point
}
\author{
Ariana Strandburg-Peshkin (primary author)

NOT YET CODE REVIEWED
}
