% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nearest_neighbor_distance.R
\name{get_nearest_neighbor_distance}
\alias{get_nearest_neighbor_distance}
\title{Get nearest neighbor distance}
\usage{
get_nearest_neighbor_distance(xs, ys, i)
}
\arguments{
\item{xs}{\verb{N x n.times} matrix giving x positions (or UTM eastings) of all \code{N} individuals over \code{n_times} timesteps}

\item{ys}{\verb{N x n.times} matrix giving y positions (or UTM northings) of all \code{N} individuals over \code{n_times} timesteps}

\item{i}{the focal individual (whose nearest neighbor distance will be computed)}
}
\value{
Returns a vector of nearest neighbor distances at each time point (length \code{n_times})
}
\description{
Gets the distance to an individual i's nearest neighbor over time
}
\author{
Ariana Strandburg-Peshkin

NOT YET CODE REVIEWED
}
