% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_positions_relative_to_group.R
\name{get_positions_relative_to_group}
\alias{get_positions_relative_to_group}
\title{Get positions relative to group}
\usage{
get_positions_relative_to_group(
  xs,
  ys,
  heading_type,
  spatial_R = NULL,
  t_window = NULL,
  forward = T,
  min_inds_tracked = NULL
)
}
\arguments{
\item{xs}{\verb{N x n_times} matrix giving x coordinates of each individual over time}

\item{ys}{\verb{N x n_times} matrix giving y coordinates of each individual over time}

\item{heading_type}{character string specifying heading type - \code{'spatial'} or \code{'temporal'}}

\item{spatial_R}{radius to use for spatial headings (if \code{heading_type = 'spatial'})}

\item{t_window}{temporal window to use for temporal headings (if \code{heading_type = 'temporal'})}

\item{forward}{whether to compute headings into the future (\code{forward = T}) or the past (\code{forward = F})}

\item{min_inds_tracked}{if specified, sets a minimum number of individuals that must be tracked at any moment in time to compute heading (otherwise all positions will be NA at that time point}
}
\value{
Returns a list containing rel_xs and rel_ys matrices of each individual's position relative to the group centroid and heading
}
\description{
Get individual positions relative to the group centroid and heading.
The group centroid is defined as the origin (0,0) and the heading points along the positive x axis.
}
\author{
Ariana Strandburg-Peshkin (primary author)

NOT YET CODE REVIEWED
}
