% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pulls_and_anchors.R
\name{get_pulls_and_anchors}
\alias{get_pulls_and_anchors}
\title{Get pulls and anchors}
\usage{
get_pulls_and_anchors(
  xa,
  xb,
  ya,
  yb,
  a,
  b,
  noise_thresh = 5,
  plot_results = F,
  include_initial_fusion = F,
  include_final_fission = F
)
}
\arguments{
\item{xa:}{x coordinates for individual a (vector of length \code{n_times})- must be continuous data (no NAs)}

\item{xb:}{x coordinates for individual b (vector of length \code{n_times})- must be continuous data (no NAs)}

\item{ya:}{x coordinates for individual a (vector of length \code{n_times})- must be continuous data (no NAs)}

\item{yb:}{x coordinates for individual b (vector of length \code{n_times})- must be continuous data (no NAs)}

\item{a:}{index of the first individual}

\item{b:}{index of the second individual}

\item{noise_thresh:}{noise threshold (defaults to 5 m)}

\item{plot_results:}{whether to plot results or not}

\item{include_initial_fusion:}{if T, the function will also output an initial fusion event. In most use cases, this should be set to false. See below for details.}

\item{include_final_fission:}{if T, the funciton will also output a final fission event. In most use cases, this should be set to false. See below for details.}
}
\value{
Returns a data frame  containing dyadic interactions between a and b.
Contains columns: \code{t1}, \code{t2}, \code{t3}, \code{initiator}, \code{responder}, \code{type}, \code{disparity}, \code{strength},\code{disparity_additive}, and \code{strength_additive}.
The \code{disparity} and \code{strength} are as defined in Strandburg-Peshkin et al. 2015, whereas
\code{disparity_additive} and \code{strength_additive} are alternative formulations of these metrics that add the components together instead of multiplying them.
}
\description{
Get successful ('pull') and failed ('anchor') dyadic interactions between individuals a and b.
}
\section{Details on fission and fusion events}{


If \code{include_initial_fusion} is set to T, the script will also return the
"fusion" defined by the beginning of the sequence and the first minimum.
If \code{include_final_fission} is set to T, the script will also return the
"fission" defined by the last minimum and the end of the sequence. If the first or
last min/max values are not minima, half events will not be returned for them.
The purpose of this option is that if this code is used on a sequence where two
individuals come together and then split apart (in fission-fusion analyses), the
first and last "half events" represent fusion and fission events, so these
are extracted. In normal use cases, this value should be set to F (default).
However in cases where this analysis is being run on the "together periods"
of two individuals in a fission-fusion analysis, these values can be set to T to return
the times and initiators of the initial fusion and eventual fission event.
In the case of fission and fusion events, there are only two time points return (\code{t1} and \code{t2}).
\code{t3} is defined as \code{NA}. The \code{initiator} is defined as the individual with the greater
displacement during the period \code{t1} to \code{t2} and the \code{responder} is the other
individual. The \code{strength} and \code{disparity} are defined in parallel to those for
pulls and anchors, but using only the one time period (\code{t1} to \code{t2}). To make the
values somewhat comparable to pulls and anchors, these values are squared.
}

\author{
Ariana Strandburg-Peshkin (primary author)

NOT YET CODE REVIEWED
}
