% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pulls_and_anchors_with_NA.R
\name{get_pulls_and_anchors_with_NA}
\alias{get_pulls_and_anchors_with_NA}
\title{Wrapper for \code{get_pulls_and_anchors} function, handling datasets with NAs}
\usage{
get_pulls_and_anchors_with_NA(
  xa,
  xb,
  ya,
  yb,
  a,
  b,
  noise_thresh = 5,
  plot_results = F,
  include_initial_fusion = F,
  include_final_fission = F,
  NA_tolerance = 0,
  min_time = NULL,
  verbose = FALSE
)
}
\arguments{
\item{xa:}{x coordinates for individual a (vector of length \code{n_times})}

\item{xb:}{x coordinates for individual b (vector of length \code{n_times})}

\item{ya:}{x coordinates for individual a (vector of length \code{n_times})}

\item{yb:}{x coordinates for individual b (vector of length \code{n_times})}

\item{a:}{index of the first individual}

\item{b:}{index of the second individual}

\item{NA_tolerance:}{How big of a gap in the vector (missing timesteps), created by NAs, is tolerated and therefore still combined into a single continuous dataset (numeric, positive integer)}

\item{min_time:}{the minimum length (timesteps) of a resulting time block that will still be used for analysis (numeric, positive integer)}

\item{verbose:}{If verbose = TRUE, then print out the size of each NA gap

For following parameters and more information, see documentation of \code{get_pulls_and_anchors} function:}

\item{noise_thresh:}{see documentation of \code{get_pulls_and_anchors} function}

\item{plot_results:}{see documentation of \code{get_pulls_and_anchors} function}

\item{include_initial_fusion:}{see documentation of \code{get_pulls_and_anchors} function}

\item{include_final_fission:}{see documentation of \code{get_pulls_and_anchors} function}
}
\value{
Returns a data frame  containing dyadic interactions between a and b.
Contains columns: \code{t1}, \code{t2}, \code{t3}, \code{initiator}, \code{responder}, \code{type}, \code{disparity}, \code{strength},\code{disparity_additive}, and \code{strength_additive}.
}
\description{
Create a list of time blocks using NAs in the dataset as seperation points, then run the \code{get_pulls_and_anchors} function on each of these blocks separately.
}
\details{
Time blocks are created by splitting the dataset at the indexes of NAs. If the length of the NA-chain is below "NA_tolerance", then the script will ignore these NAs and not split at that position.
Instead, these NAs will be filled in with the value of the next (non-NA) position in the vector. TODO: check this - this seems to be what it's doing, but is this correct and is this the desired behavior?
This results in the datapoints staying connected/in a single time block for small breaks. If the NA-chain in the dataset is longer than "NA_tolerance",
then the values before and after that chain are split into separate time blocks.
If "min_time" != NULL, then the script will remove all time blocks shorter than the specified value (usually time steps), resulting only in time blocks of relevant lengths.

Note that if either \code{include_initial_fusion} or \code{include_final_fission} are true, this function will treat each time block separately. In other words,
a fission and/or fusion might be detected for each time block. TODO: I added this. Please check for correctness.

If "Error in xa_blocks[\link{i}] : subscript out of bounds" appears, then the min_time set excludes all time blocks, leaving no events to analyse.
TODO: We might want to check this within the function and have it return NULL instead of throwing an error. This is what the main function does if there are no events.
}
\author{
Dario Walser

NOT YET CODE REVIEWED
}
