% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_simplified_turn_and_speed_influence.R
\name{get_simplified_turn_and_speed_influence}
\alias{get_simplified_turn_and_speed_influence}
\title{Get simplified turn and speed influence}
\usage{
get_simplified_turn_and_speed_influence(
  xs,
  ys,
  heading_type,
  idx_breaks = NULL,
  spatial_R = NULL,
  t_window = NULL,
  min_percentile = 0.5,
  seconds_per_time_step = 1,
  min_left_speed = NULL,
  min_right_speed = NULL,
  min_left_pos = NULL,
  min_right_pos = NULL,
  min_faster_speed_diff = NULL,
  min_slower_speed_diff = NULL,
  min_front_pos = NULL,
  min_back_pos = NULL,
  verbose = T,
  output_details = T
)
}
\arguments{
\item{xs}{\verb{N x n_times} matrix giving x coordinates of each individual over time}

\item{ys}{\verb{N x n_times} matrix giving y coordinates of each individual over time}

\item{heading_type}{character string specifying heading type - \code{'spatial'} or \code{'temporal'}}

\item{idx_breaks}{vector of indexes to breaks in the data (e.g. breaks between days)}

\item{spatial_R}{radius to use for spatial headings (if \code{heading_type = 'spatial'})}

\item{t_window}{temporal window to use for temporal headings (if \code{heading_type = 'temporal'})}

\item{min_percentile}{minimum percentile to use for left / right relative speed or distance}

\item{seconds_per_time_step}{number of seconds corresponding to each time step}

\item{min_left_speed}{minimum speed when i moving to the left to include data (positive value)}

\item{min_right_speed}{minimum speed when i moving to the right to include data (positive value)}

\item{min_left_pos}{minimum distance of i to the left of j to include data (positive value)}

\item{min_right_pos}{minimum distance of i to the right to j include data (positive value)}

\item{min_faster_speed_diff}{minimum speed difference between i and j when i is going faster to include data (positive value)}

\item{min_slower_speed_diff}{minimum speed difference between i and j when i is going slower to include data (positive value)}

\item{min_front_pos}{minimum position toward the front when i is in front of j to include data (positive value)}

\item{min_back_pos}{minimum position toward the back when i is in back of j to include data (positive value)}

\item{verbose}{whether the print progress (if \code{T})}

\item{centroid}{whether to use the group centroid (if \code{centroid = T}) instead of computing influence for each dyad (if \code{centroid = F})}

\item{symmetrize_lr}{whether to symmetrize the thresholds for right and left movement + position (default \code{T})}
}
\value{
Returns an \verb{N x N} matrix of the turn influence of individual \code{i} (row) on individual \code{j} (column).
}
\description{
Get the simplified turn influence between all pairs of individuals
TODO: Update documentation to incorporate speed influence stuff
\code{i} and \code{j}, or, if \code{centroid = T} for all indiividuals \code{i} on the group centroid.
The \verb{simplified movement turn influence} of individual \code{i} on individual \code{j}
is defined as the probability that \code{j} turns right in the future given that \code{i} moved
right relative to \code{j}'s past heading at a speed above a threshold value
\code{min_right_speed} or left relative to \code{j}'s past heading at a speed above a threshold
value \code{min_left_speed}. This will be computed if \code{influence_type} is set to \code{'movement'}.
The \verb{simplified positional turn influence} of individual \code{i} on individual \code{j}
is defined as the probability that \code{j} turns right in the future given that \code{i} was located
at least \code{min_right_dist} meters right of individual \code{j} relative to \code{j}'s current position
and heading or the probability that \code{j} turns left in the future given that \code{i} was located
at least \code{min_left_dist} meters left of individual \code{j} relative to \code{j}'s current position
and heading. This will be computed if \code{influence_type} is set to \code{position}.
The parameters \code{min_right_dist} and \code{min_left_dist}, or \code{min_right_speed} and
\code{min_left_speed} are either set manually or, if \code{min_percentile} is specified, based on percentiles
of the left-right relative speed and front-back relatives speed distributions across all dyads
(or across all individuals relative to the group centroid, if \code{centroid = T}). The variable
\code{min_percentile} overrides the manually set minimum speeds / distances.
}
\details{
The minimum speeds (or distances) are computed as follows. First, split the distribution of relative left-right
speeds (or distances) across all dyads into those greater than and those less than 0. Then, compute the quantile
\code{min_percentile} for each distribution separately. Finally, set \code{min_left_speed} and \code{min_right_speed} at the computed
values. If \code{symmetrize_lr = T}, take instead the mean absolute value of the quantiles for both distributions and use it
for both thresholds (this will result in using the same relative speed when going right or left as a threshold - otherwise these could be different if \code{symmetrize = F})

The same can be done for the relative left and right distance distributions to set
\code{min_left_dist} and \code{min_right_dist} respectively, if \code{influence_type = 'position'}
}
\author{
Ariana Strandburg-Peshkin

NOT YET CODE REVIEWED
}
