% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spatially_discretized_trajectories.R
\name{get_spatially_discretized_trajectories}
\alias{get_spatially_discretized_trajectories}
\title{Get spatially discretized trajectories}
\usage{
get_spatially_discretized_trajectories(xs, ys, R, breaks = NULL, verbose = T)
}
\arguments{
\item{xs}{\verb{N x n_times} matrix giving x coordinates of each individual over time}

\item{ys}{\verb{N x n_times} matrix giving y coordinates of each individual over time}

\item{R}{radius (ruler length) used in spatial discretization}

\item{breaks}{vector of indexes to breaks in the data (e.g. breaks between days)}

\item{verbose}{whether to print progress while running}
}
\value{
Returns a list containing \code{spat_ts} (the time points associated with each point along the spatially discretized trajectory),
\code{spat_xs} (the x coordinates of each point along the spatially discretized trajectory),
\code{spat_ys} (the y coordinates of each point along the spatially discretized trajectory),
\code{spat_breaks} (indexes to starts of breaks in the new spatially discretized data),
\code{R} (radius used)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Gets spatially discretized trajectories from temporally discretized trajectories,
using a certain "ruler length" R.

TODO: Code doesn't deal well with strings of NAs within it - look into this.
}
\author{
Ariana Strandburg-Peshkin (primary author)

NOT YET CODE REVIEWED
}
