% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_speed_over_time.R
\name{get_speed_over_time}
\alias{get_speed_over_time}
\title{Get speed over time}
\usage{
get_speed_over_time(x_i, y_i, time_res, step = 1)
}
\arguments{
\item{x_i}{x coordinates of the individual (a vector whose length is the number of timesteps) or of a group centroid}

\item{y_i}{y coordinates of the individual (a vector whose length is the number of timesteps) or of the group centroid}

\item{time_res}{the number of seconds between timestamps (e.g. for data recorded at 1 Hz, the value would be 1, for data recorded every 30 sec, the value would be 30)}

\item{step}{the time step over which to compute speed (number of time steps)}
}
\value{
Returns \code{speeds}: a time series of the speed of individual
}
\description{
Gets an individual's instantaneous speed over time, using a specified time window to compute the speed
}
\details{
The speed is computed as \code{sqrt ( (x.i[t+step] - x.i[t] )^2 + ( y.i[t+step] - y.i[t] )^2 ) / (dt * step)}
}
