% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_turn_and_speed_influence_simplified.R
\name{get_turn_and_speed_influence_simplified}
\alias{get_turn_and_speed_influence_simplified}
\title{Get simplified turn and speed influence}
\usage{
get_turn_and_speed_influence_simplified(
  xs,
  ys,
  heading_type,
  centroid = F,
  breaks = NULL,
  spatial_R = NULL,
  t_window = NULL,
  min_percentile = 0.5,
  seconds_per_time_step = 1,
  symmetrize_lr = T,
  min_left_speed = NULL,
  min_right_speed = NULL,
  min_left_pos = NULL,
  min_right_pos = NULL,
  min_faster_speed_diff = NULL,
  min_slower_speed_diff = NULL,
  min_front_pos = NULL,
  min_back_pos = NULL,
  verbose = T,
  output_details = T
)
}
\arguments{
\item{xs}{\verb{N x n_times} matrix giving x coordinates of each individual over time}

\item{ys}{\verb{N x n_times} matrix giving y coordinates of each individual over time}

\item{heading_type}{character string specifying heading type - \code{'spatial'} or \code{'temporal'}}

\item{centroid}{whether to use the group centroid (if \code{centroid = T}) instead of computing influence for each dyad (if \code{centroid = F})}

\item{breaks}{vector of indexes to breaks in the data (e.g. breaks between days)}

\item{spatial_R}{radius to use for spatial headings (if \code{heading_type = 'spatial'})}

\item{t_window}{temporal window to use for temporal headings (if \code{heading_type = 'temporal'})}

\item{min_percentile}{minimum percentile to use for left / right relative speed or distance}

\item{seconds_per_time_step}{number of seconds corresponding to each time step}

\item{symmetrize_lr}{whether to symmetrize the thresholds for right and left movement + position (default \code{T})}

\item{min_left_speed}{minimum speed when i moving to the left to include data (positive value)}

\item{min_right_speed}{minimum speed when i moving to the right to include data (positive value)}

\item{min_left_pos}{minimum distance of i to the left of j to include data (positive value)}

\item{min_right_pos}{minimum distance of i to the right to j include data (positive value)}

\item{min_faster_speed_diff}{minimum speed difference between i and j when i is going faster to include data (positive value)}

\item{min_slower_speed_diff}{minimum speed difference between i and j when i is going slower to include data (positive value)}

\item{min_front_pos}{minimum position toward the front when i is in front of j to include data (positive value)}

\item{min_back_pos}{minimum position toward the back when i is in back of j to include data (positive value)}

\item{verbose}{whether the print progress (if \code{T})}

\item{output_details}{whether to output computed matrices and arrays used in influence computations (if \code{T})}
}
\value{
Returns a list containing matrices for the 4 different types of influence, as well as other computed metrics (if \code{output_details = T}).
See "Additional information about returned objects" section below for a list of outputs.
}
\description{
Get the simplified turn and speed influence (movement and positional) between all pairs of individuals
\code{i} and \code{j}, or, if \code{centroid = T} for all indiividuals \code{i} on the group centroid excluding \code{i}.
}
\details{
The simplified movement turn influence of individual \code{i} on individual \code{j} (\code{turn_influence_movement[i,j]})
is defined as the probability that \code{j} turns right in the future given that \code{i} moved
right relative to \code{j}'s past heading at a speed above a threshold value \code{min_right_speed},
that \code{j} turns left in the future given that \code{i} moved left relative to \code{j}'s past heading at a speed above a threshold
value \code{min_left_speed}.

The simplified positional turn influence of individual \code{i} on individual \code{j} (\code{turn_influence_position[i,j]})
is defined as the probability that \code{j} turns right in the future given that \code{i} was located
at least \code{min_right_dist} meters right of individual \code{j} relative to \code{j}'s current position
and heading, or that \code{j} turns left in the future given that \code{i} was located
at least \code{min_left_dist} meters left of individual \code{j} relative to \code{j}'s current position
and heading.

The simplified movement speed influence of individual \code{i} on individual \code{j} (\code{speed_influence_movement[i,j]})
is defined as the probability that \code{j} speeds up in the future along the vector defined by \code{j}'s heading in the past,
given that \code{i} was moving at least a speed \code{min_faster_speed_diff} faster than \code{j} along \code{j}'s past heading direction, or
slows down along the vector defined by \code{j}'s heading in the past given that \code{j} was moving at least a speed \code{min_slower_speed_diff}
slower than \code{i} along the same vector in the past.

The simplified positional speed influence of individual \code{i} on individual \code{j} (\code{speed_influence_position[i,j]})
is defined as the probability that \code{j} speeds up in the future along the vector defined by \code{j}'s heading in the past
given that \code{i} was in front of \code{j} by at least \code{min_front_pos} meters relative to \code{j}'s current position and past heading,
or that \code{j} slows down in the future along the vector defined by \code{j}'s heading in the past given that \code{i} was
behind \code{j} by at least \code{min_back_pos} relative to \code{j}'s current position and past heading.

Future headings and speeds are computed either using a fixed temporal window \code{t_window} forward in time (if \code{heading_type = 'temporal'})
or based on spatial discretization with a radius \code{spatial_R} (if \code{heading_type = 'spatial'}), i.e. defining the heading based on the
when the focal individual has moved a distance \code{spatial_R} from its current location. Similarly, past headings and speeds
are computed by looking backward in time either at a fixed temporal window (\code{t_window}) or based on a fixed displacement distance (\code{spatial_R}).

The parameters specifying minimum speeds and distances left-right and front-back are either set manually (not recommended) or,
if \code{min_percentile} is specified, based on percentiles of the relevant distributions of these variables across all dyads
(or across all individuals relative to the group centroid, if \code{centroid = T}). The variable
\code{min_percentile} overrides the manually set minimum speeds / distances.
The minimum speeds (or distances) are computed as follows. First, split the distribution of relative left-right
speeds (for example) across all dyads into those greater than and those less than 0. Take the absolute value of all speeds to
define a positive speed for both directions. Next, compute the quantile \code{min_percentile} for each distribution (left and right) separately.
Finally, set \code{min_left_speed} and \code{min_right_speed} at the computed
values. If \code{symmetrize_lr = T} (recommended), take instead the mean value of the quantiles for both distributions and use it
for both thresholds (this will result in using the same relative speed when going right or left as a threshold - otherwise these could be different if \code{symmetrize_lr = F})
The same procedure is done for left-right positions, front-back relative speeds, and front-back positions, using the same value of \code{min_percentile}. In the case of
speeds, no symmetrization is done because speed differences and front-back distances are not expected to be symmetrical.

If \code{breaks} is set, headings that go across breaks in the data as specified by this vector are not included in the computations.
}
\section{Additional information about returned objects}{


The output list conatins the following objects:

\code{params} list of parameters used in the calculation (see input variables for descriptions)

\code{turn_influence_movement}: \verb{i x j} matrix representing movement turn influence of individual \code{i} on individual \code{j}

\code{turn_influence_position}: \verb{i x j} matrix representing positional turn influence of individual \code{i} on individual \code{j}

\code{speed_influence_movement}: \verb{i x j} matrix representing movement speed influence of individual \code{i} on individual \code{j}

\code{speed_influence_position}: \verb{i x j} matrix representing positions speed influence of individual \code{i} on individual \code{j}

If \code{centroid = T}, everything is instead computed at the centroid level (with the centroid as follower, replacing individual \code{j}).
In this case, the output matrices will still be the same shape as those for the dyadic case, but they will have \code{NA}s for all columns
except \code{j=1} (representing the influence of each individual \code{i} on the centroid excluding that individual).
This parallel structure is for compatibility between dyadic and centroid-based analyses.

If \code{output_details = T}, some additional parameters and the following computed matrices / arrays will also be outputted:

\code{lr_speed}: \verb{i x j x t} array of the past left-right speed of individual \code{i} relative to the past heading of individual \code{j} at time \code{t}

\code{lr_pos}: \verb{i x j x t} array of the past left-right position of individual \code{i} relative to the current position and past heading of individual \code{j} at time \code{t}

\code{fb_speed}: \verb{i x j x t} array of the past front-back speed difference between individual \code{i} and individual \code{j} along the vector of the past heading of individual \code{j} at time \code{t}

\code{fb_pos}: \verb{i x j x t} array of the past front-back position of individual \code{i} relative to the current position and past heading of individual \code{j} at time \code{t}

\code{turn_angle}: \verb{j x t} matrix of the turning angle of individual \code{j} between the past and future at time \code{t}

\code{speed_up}: \verb{j x t} matrix of the difference in speed of individual \code{j} between future and past at time \code{t}, project along \code{j}'s own past heading

The output matrices \code{turn_angle[j,t]} and \code{speed_up[j,t]} then represent the centroid excluding individual \code{j}, rather than
the data for individual \code{j} as in the dyadic case.
Similarly, the output arrays with dimensions \verb{i x j x t} will have \code{NA}s for all values of \code{j > 1}, and the \code{j = 1}
values represent the speeds / positions of individuals relative to the centroid (excluding \code{i})
}

\author{
Ariana Strandburg-Peshkin

NOT YET CODE REVIEWED
}
