% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_gps_level0_to_level1.R
\name{gps_preprocess_level0_to_level1}
\alias{gps_preprocess_level0_to_level1}
\title{Preprocess GPS data (in matrix format) from level 0 to level 1}
\usage{
gps_preprocess_level0_to_level1(
  input_file_path = NULL,
  output_file_path = NULL,
  xs = NULL,
  ys = NULL,
  timestamps = NULL,
  ids = NULL,
  remove_unrealistic_speeds = T,
  remove_isolated_points = T,
  remove_unrealistic_locations = T,
  interpolate_small_gaps = T,
  interpolate_stationary_periods = T,
  max_speed_percentile = 0.9995,
  max_speed = NULL,
  max_sd_away = 10,
  max_dist_percentile = 0.9999,
  max_isolated_point_dist = 1000,
  max_interp_len = 5,
  max_move_dist = 5,
  max_move_time = 5,
  verbose = T
)
}
\arguments{
\item{input_file_path}{full path to the input file containing \code{xs}, \code{ys}, \code{timestamps}, and \code{ids} (overrides manual passing in of these parameters), must be an RData file}

\item{output_file_path}{full path to the output file where the level 1 dataset will be stored, must end in .RData}

\item{xs}{\verb{N x n_times} matrix giving x coordinates of each individual over time (if an input file is not specified, pass this in manually)}

\item{ys}{\verb{N x n_times} matrix giving y coordinates of each individual over time (if an input file is not specified, pass this in manually)}

\item{timestamps}{vector of timestamps (if an input file is not specified, pass this in manually)}

\item{ids}{data frame containing information about each individual (if an input file is not specified, pass this in manually)}

\item{remove_unrealistic_speeds}{whether to remove unrealistic speeds (\code{T} or \code{F})}

\item{remove_isolated_points}{whether to remove isolated points (\code{T} or \code{F})}

\item{remove_unrealistic_locations}{whether to remove unrealistic locations (\code{T} or \code{F})}

\item{interpolate_small_gaps}{whether to interpolate small gaps (\code{T} or \code{F})}

\item{interpolate_stationary_periods}{whether to interpolate stationary periods (\code{T} or \code{F}) - cannot be run unless \code{interpolate_small_gaps} is also \code{T}}

\item{max_speed_percentile}{quantile to use to define the maximum speed}

\item{max_speed}{maximum speed (overrides max_speed_percentile if specified)}

\item{max_sd_away}{standard deviation of xs and ys distributions for each individual beyond which points will be removed}

\item{max_dist_percentile}{quantile to use to define the maximum x and y coordinates (all those outside will be removed)}

\item{max_isolated_point_dist}{maximum isolated point distance}

\item{max_interp_len}{maximum length of an \code{NA} gap to linearly interpolate (number of time points)}

\item{max_move_dist}{maximum distance moved during a time \code{max_move_time} to interpolate using the average position}

\item{max_move_time}{maximum time of a gap to interpolate if stationary (in timesteps)}

\item{verbose}{whether to print out progress and information}
}
\value{
Returns a list containing new \code{xs} and \code{ys} matrices, and also saves them plus the \code{timestamps} and \code{ids} objects to an output file if specified
}
\description{
This function takes in a level 0 dataset (raw GPS data) and performs minimal
pre-processing to return a level 1 dataset.
The function performs the following steps (in order):
}
\details{
\enumerate{
\item If \code{remove_unrealistic_speeds = T}, removes unrealistic speeds (greater than \code{max_speed_percentile}) and replaces them with NAs (default .9995 quantile, or alternatively a max speed \code{max_speed} can be set manually.
\item If \code{remove_isolated_points = T}, finds extreme distances \verb{> max_dist_percentile} quantile (default 99.99\%) or \verb{< 1 - max_dist_percentile} of \code{xs} or \code{ys} for each individual and, if there are no other points from that individual within \code{max_isolated_point_dist} (default 1000 m) of that point, replaces them with \code{NA}s
\item If \code{remove_unrealistic_locations = T}, finds extreme xs and ys above \code{mean + sd * max_sd_away} (default 10) for each ind and removes those
\item If \code{interpolate_small_gaps = T}, fills in missing data gaps less than length \code{max_interp_len} with linear interpolation (default 5)
\item If \code{interpolate_stationary_periods = T}, finds instances where an animal did not move more than \code{max_move_dist} (default 5 m) during an \code{NA} gap of \verb{< max_move_time} (default 300 timesteps) and replaces
them with the mean location of the individual between start and end of the sequence
}
}
