% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_splits_and_merges.R
\name{identify_splits_and_merges}
\alias{identify_splits_and_merges}
\title{Identify group splits and merges from multi-individual trajectory data}
\usage{
identify_splits_and_merges(
  R_inner,
  R_outer,
  xs,
  ys,
  ts,
  breaks = c(1, length(ts) + 1),
  names = NULL,
  break_by_day = F,
  verbose = T
)
}
\arguments{
\item{R_inner}{inner distance threshold to identify periods of connectedness (numeric)}

\item{R_outer}{outer distance threshold to identify periods of connectedness (numeric)}

\item{xs}{UTM eastings matrix (\code{n_ninds} x \code{n_times} matrix where xs\link{i,t} gives the easting of individual i at time step t)}

\item{ys}{UTM northing matrix (\code{n_ninds} x \code{n_times} matrix where ys\link{i,t} gives the northing of individual i at time step t)}

\item{ts}{vector of timestamps (POSIXct), must have same dimensions as columns of \code{xs} and \code{ys} matrices}

\item{breaks}{indexes to breaks in the data (default NULL treats data as a contiguous sequence). If specified, overrides \code{break_by_day}}

\item{break_by_day}{whether to break up data by date (T or F)}
}
\value{
a list containing:

\code{events_detected}: data frame with info on detected fissions and fusions, and limited info for shuffles, including columns...
\code{events_detected$event_idx}: unique id number of the event
\code{events_detected$tidx}: (initial) time index of the event
\code{events_detected$event_type}: "fission" or "fusion" or "shuffle
\code{events_detected$n_groups_before}: number of groups prior to the event
\code{events_detected$n_groups_after}: number of groups after the event
\code{events_detected$big_group_idxs}: indexes of all the individuals involved in the event
\code{events_detected$big_group}: names of all the individuals involved in the event
\code{events_detected$group_A_idxs}, $group_B_idxs, $group_C_idxs, etc.: individual idxs of subgroup members
\code{events_detected$group_A}, \verb{$group_B}, \verb{$group_C}, etc.: names of subgroup members
\code{events_detected$n_A}, \verb{$n_B}, \verb{$n_C} etc.: number of individuals in each subgroup
\code{events_detected$n_big_group}: number of individuals in the big group (original group for fissions, subseq group for fusions)

(NOTE: \code{big_group_idxs}, \code{big_group}, \code{group_A_idxs} etc.,
\code{group_A} etc. \code{n_A} etc. and \code{n_big_group} are set to NA for shuffles...
you can get more detailed info for shuffles from \code{all_events_info} object))

\code{all_events_info}: list of information about all fission-fusion (or shuffle)
\verb{events. all_events_info[[i]]} contains the following info for event i:
\code{all_events_info[[i]]$t}: time index of the event
\code{all_events_info[[i]]$groups_before}: (list of lists) list of groups before the event (at time t)
\code{all_events_info[[i]]$groups_after}: (list of lists) list of groups after the event (at time t + 1)
\verb{all_events_info[[i]]event_type}: 'fission', 'fusion', or 'shuffle' (character string)
\code{all_events_info[[i]]$n_groups_before}: number of groups before the event
\code{all_events_info[[i]]$n_groups_after}: number of groups after the event

\code{groups_list}: list of subgroups in each timestep
\code{groups_list[[t]]} gives a list of the subgroups
\code{groups_list[[t]][[1]]} gives the vector of the first subgroup, etc.

\code{together}: \link{n_inds x n_inds x n_times array} of whether dyads are
connected (1) or not (0) or unknown (NA)

\code{R_inner}: inner radius used in the computations (same as \code{R_inner} above)

\code{R_outer}: outer radius used in the computations (same as \code{R_outer} above)
}
\description{
Detects splits and merges (a.k.a. fissions and fusions) using "sticky-DBSCAN"
method from Libera et al. 2023.
}
