% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_axytrek_gps_file.R
\name{import_axytrek_gps_file}
\alias{import_axytrek_gps_file}
\title{Import Axy-Trek GPS file}
\usage{
import_axytrek_gps_file(input_file_path)
}
\arguments{
\item{input_file_path}{full path to the input file}
}
\value{
Returns a data frame with the following columns: \code{timestamp}, \code{location.lat},
\code{location.long}, \code{V4}, \code{V5}, \code{satellite.count}, \code{V7}
}
\description{
Imports GPS data from an Axy Trek file, which may have lines that are not
formatted such that read.delim can parse them. Can currently handle files
from 2021-2023 at least. Output data frame has column names compatible with
the function \code{cocomo::reformat_movebank_to_matrix}.
}
\author{
Ariana Strandburg-Peshkin (primary author)

NOT YET CODE REVIEWED
}
