% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_behavioral_seqs_to_matrix.R
\name{import_behavioral_seqs_to_matrix}
\alias{import_behavioral_seqs_to_matrix}
\title{Import behavioral sequences (output from classifier) to matrix format}
\usage{
import_behavioral_seqs_to_matrix(input_dirs, ids, timestamps)
}
\arguments{
\item{input_dirs}{list of input directories where behavioral sequence data are stored (full paths)}

\item{ids}{data frame giving individual ids for that deployment}

\item{timestamps}{vector of timestamps associated with that deployment (in UTC)}
}
\value{
Returns a list containing \code{out$behavs}, a matrix where \code{out$behavs[i,t]} gives the behavior of
individual \code{i} (correspoding to the row in \code{ids} at time \code{t} (corresponding to the index in \code{timestamps})
In this matrix, the behavior is represented as an integer value (for quicker processing). The second item
in the list, \code{out$behavs_key} is a data frame with columns \code{behav} (the behavior as a string) and \code{int_value}
(the correspoding integer value associated with that behavior in the \code{out$behavs} matrix)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Imports behavioral sequence data from a set of csv files and converts them to
matrix format matching the \code{xs} and \code{ys} matrices used in the \code{cocomo} library
}
\author{
Ariana Strandburg-Peshkin (primary author)

NOT YET CODE REVIEWED
}
