% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_meerkat_gps_data.R
\name{import_meerkat_gps_data}
\alias{import_meerkat_gps_data}
\title{Import meerkat data to matrix format}
\usage{
import_meerkat_gps_data(input_dir, tag_type = "gipsy5", timezone = "UTC")
}
\arguments{
\item{input_dir}{full path to input directory where all files from a given deployment are stored (e.g. \code{"~/EAS_shared/meerkat/archive/rawdata/meerkat_movecomm_2017/HM_2017_1/"})}

\item{tag_type}{\code{'gipsy5'} or \code{'axytrek'}}

\item{data_chunks}{data frame with columns \code{start} and \code{end} specifying starting and ending datetimes (in POSIXct format - with time zone!) for each contiguous chunk of data. overrides other specified start and end times / dates.}

\item{seconds_per_time_step}{sampling interval of GPS fixes (in seconds)}

\item{start_date}{a character string specifying the starting date for data sampled in a daily basis, only used if \code{data_chunks = NULL}, format must be \code{'YYYY-MM-DD'}}

\item{end_date}{a character string specifying the end date for data sampled in a daily basis, only used if \code{data_chunks = NULL}, format must be \code{'YYYY-MM-DD'}}

\item{start_time}{a character string specifying the start time for data sampled on a daily basis, only used if \code{data_chunks = NULL}, format must be \code{'HH:MM:SS'}}

\item{end_time}{a character string specifying the end time for data sampled on a daily basis, only used if \code{data_chunks = NULL}, format must be \code{'HH:MM:SS'}}

\item{utm_zone}{numeric UTM zone (only needed if \code{output_utm = T} and if \code{movebank_data} does not have \code{utm_easting} and \code{utm_northing} columns)}

\item{hemisphere}{hemisphere (\code{'north'} or \code{'south'}) for UTM calculations (only needed if \code{output_utm = T} and if \code{movebank_data} does not have \code{utm_easting} and \code{utm_northing} columns)}
}
\description{
Import meerkat data to matrix format
}
