% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_format_to_df.R
\name{matrix_format_to_df}
\alias{matrix_format_to_df}
\title{Convert from matrix format to dataframe}
\usage{
matrix_format_to_df(
  xs,
  ys,
  timestamps,
  ids = NULL,
  time_indices_selected,
  ids_selected = NULL,
  lats = NULL,
  lons = NULL
)
}
\arguments{
\item{xs}{UTM eastings matrix (\code{n_inds} x \code{n_times} matrix where xs\link{i,t} gives the easting of individual i at time step t)}

\item{ys}{UTM northings matrix (\code{n_inds} x \code{n_times} matrix where ys\link{i,t} gives the northing of individual i at time step t)}

\item{timestamps}{vector of timestamps (POSIXct), must have same dimensions as columns of \code{xs} and \code{ys} matrices}

\item{ids}{Vector of individuals ordered the same as the rows of xs and ys.}

\item{time_indices_selected}{Time indices for data to include in the dataframe, to be matched with xs, ys, timestamps}

\item{ids_selected}{Identities of individuals to include in dataframe. If none provided, all individuals are included}

\item{lats}{matrix of latitude values (\code{n_inds} x \code{n_times})}

\item{lons}{matrix of longitude values (\code{n_inds} x \code{n_times})}
}
\value{
Returns a dataframe of ids, locations, and times
}
\description{
Creates a dataframe out of GPS data in "matrix format" suitable for plotting with ggplot
}
\author{
Eli Strauss

NOT YET CODE REVIEWED
}
