% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meerkat_auto_label_synch_times.R
\name{meerkat_auto_label_synch_times}
\alias{meerkat_auto_label_synch_times}
\title{Auto label synch times
Automatic labeling of synch calls that are not labeled with a time, assuming they should be close to multiples of 90 sec apart
File must have at least one fully labeled synch - and make sure this is correct!
TODO: add more details on how this works here
TODO: change max drift to be per time in file instead of absolute}
\usage{
meerkat_auto_label_synch_times(
  label_file,
  labeled_synchs,
  outdir,
  machine_labels = T,
  likeli_thresh = 0.6,
  max_drift = 1
)
}
\arguments{
\item{label_file}{path to label file}

\item{labeled_synchs}{data frame of labeled synchs associated with that file}

\item{outdir}{output directory}

\item{likeli_thresh}{likelihood threshold (default to 0.6) - only use synchs above this threshold (only relevant if machine_labels = T)}

\item{max_drift}{maximum amount of drift allowed to still consider a synch call to be correctly positioned}

\item{machine_labelswhether}{the labels are machine generated (T) or not (F), defaults to T}
}
\value{
Saves a new label file to the selected directory, with the original filename with _autosync.csv appended
Also returns the name of the output file (invisibly)
}
\description{
Auto label synch times
Automatic labeling of synch calls that are not labeled with a time, assuming they should be close to multiples of 90 sec apart
File must have at least one fully labeled synch - and make sure this is correct!
TODO: add more details on how this works here
TODO: change max drift to be per time in file instead of absolute
}
\author{
Ariana Strandburg-Peshkin

NOT YET CODE REVIEWED
}
