% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meerkat_check_data_availability_sheets.R
\name{meerkat_check_data_availability_sheet}
\alias{meerkat_check_data_availability_sheet}
\title{Check meerkat data availability sheets for any inconsistencies or errors}
\usage{
meerkat_check_data_availability_sheet(
  path_to_data_availability_sheet,
  path_to_data_dir
)
}
\arguments{
\item{path_to_data_availability_sheet}{path to data availability sheet on the server (.csv)}

\item{path_to_data_dir}{path to the outer directory of the relevant deployment}
}
\value{
0 if successful
}
\description{
Reads in a data availability sheet and checks for the following:
\enumerate{
\item Flag any missing fields
\item Check internal consistency of each row
}
\itemize{
\item datatype, tagtype, and datasource are all consistent and into expected categories
\item normalday column is either 0 or 1
\item start and end are correctly formatted times
\item filename contains individual id code
\item filename has expected file extension
\item in audio files, filename contains date matching start date
\item normalday and comments columns make sense together
}
\enumerate{
\item Check that file in filename column exists on server
\item Finds files on the server that involve an individual from the group on a
date included in the deployment, but aren't in the data availability sheet
}
}
\details{
Outputs everything into two csvs for manual checking:
\itemize{
\item \code{path_to_data_availability_sheet}_check.csv: rows to check in data availability sheet
\item \code{path_to_data_availability_sheet}_missing.csv: potential missing files (filling gaps in data availability sheet)
}
}
\author{
Ariana Strandburg-Peshkin (primary author)

NOT YET CODE REVIEWED
}
