% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meerkat_synch_audio_file_labels_to_UTC.R
\name{meerkat_synch_audio_file_labels_to_UTC}
\alias{meerkat_synch_audio_file_labels_to_UTC}
\title{Synch labels within audio file to UTC time}
\usage{
meerkat_synch_audio_file_labels_to_UTC(
  path_to_label_file,
  path_to_synch_file = "~/EAS_shared/meerkat/working/METADATA/total_synch_info.csv",
  min_offset_outlier = 0.5,
  min_n_synchs = 3,
  min_frac_spanned_by_synchs = 0.2,
  make_plot = T
)
}
\arguments{
\item{path_to_label_file}{path to the label file from Audition}

\item{path_to_synch_file}{path to the synch file}

\item{min_offset_outlier}{minimum offset (in seconds) between fitted times and labeled talking clock times to be considered an outlier}

\item{min_n_synchs}{minimum number of synchs (after excluding outliers) to perform a fit}

\item{min_frac_spanned_by_synchs}{minimum fraction of the total file length (between first and last label time) spanned by synch calls to complete the synching}

\item{make_plot}{whether to also output a plot showing the synchs in time in recording vs talking clock time, with the final fit and outliers indicated}
}
\value{
Returns a list containing \code{filename} (base name of the label file),
\code{synch_completed} (T or F whether the synch was completed successfully - if F, no other objects are returned in the list)
\code{labels_synched} (data frame with columns \code{Name},\code{duration},\code{start_UTC},\code{start_time_in_file},\code{filename}),
\code{synchs_used} (data frame containing synch labels used in the fit, and computed information about them)
and \code{outliers} (data frame with points labeled as outliers as well as filenames, plus some other columns)
}
\description{
Reads in a label file (in Audition format) and a synch file, gets synch points,
and synchs all labels in the file to UTC. Outputs a table with an additional
column specifying timestamp_UTC. Currently only designed to work with meerkat data.
}
\details{
The label file should be in the format of Audition labels. The Name column must contain
labels of the form 'synch H:MM:SS' or 'synch MM:SS' to specify the times of synch calls
as heard on the talking clock.
}
\author{
Ariana Strandburg-Peshkin (primary author)

NOT YET CODE REVIEWED
}
