% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_audition_time.R
\name{parse_audition_time}
\alias{parse_audition_time}
\title{Parse Adobe Audition formatted time}
\usage{
parse_audition_time(audition_time_str)
}
\arguments{
\item{audition_time_str}{character string correspoding to an Audition-formatted time}
}
\value{
Returns numeric value of time into the file (in seconds)
}
\description{
Parses a time from the \code{Start} or \code{Duration} column of an Audition label file to seconds
}
\details{
The \code{Start} and \code{Duration} columns of Adobe Audition label files have a non-standard
format. Within the first 10 minutes, they are formatted as M:SS.SSS. After 10 minutes
they are formatted as MM:SS.SSS and after the first hour they are formatted as
H:MM:SS.SSS (and presumably after the first 10 hours as HH:MM:SS.SSS). This function
parses these weird labels and returns a numeric value of the time into the file, in seconds
}
\author{
Ariana Strandburg-Peshkin (primary author)

NOT YET CODE REVIEWED
}
