% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_behav_and_calls.R
\name{plot_behav_and_calls}
\alias{plot_behav_and_calls}
\title{Plot individual behavior and calls during a time period specified by the user.}
\usage{
plot_behav_and_calls(
  behavs,
  calls_array,
  behavs_key,
  focal_ind,
  t0,
  tf,
  nonfocal_calls_to_plot = NULL,
  nonfocal_behavs_to_plot = NULL,
  smooth_window = 31
)
}
\arguments{
\item{behavs}{matrix of dimensions \code{n_inds} x \code{n_times} where \code{behavs[i,t]} gives the behavior (numeric) of individual \code{i} at time step \code{t}}

\item{calls_array}{array of dimensions \code{n_inds} x \code{n_times} x \code{n_calltypes} where \code{calls_array[i,t,c]} indicates the number of calls of type \code{c} given by the individual \code{i} at time step \code{t}}

\item{behavs_key}{data frame of behavior types with columns \code{behav} (behavior string) and \code{int_value} (its integer value in the \code{behavs} matrix)}

\item{focal_ind}{integer giving the index of the focal individual}

\item{t0}{integer giving the time step to start at}

\item{tf}{integer giving the time step to stop at}

\item{nonfocal_calls_to_plot}{vector of character strings indicating which calls to plot for nonfocal individuals (must match call types in \code{calls_array} 3rd dimension names)}

\item{nonfocal_behavs_to_plot}{vector of character strings indicating which behaviors to plot for nonfocal individuals (must match behavior types in \code{behavs_key})}

\item{smooth_window}{smoothing window for indicating presence of nonfocal calls (in time steps)}

\item{calls_to_plot}{vector of character strings indicating call types to plot (other call types will get lumped into an "other" category)}

\item{behavs_to_plot}{vector of character strings indicating behavior types of plot (other behavior types will get lumped into an "other" category)}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Also include a summary of the behavior of the rest of the group.
}
\author{
Ariana Strandburg-Peshkin (primary author)

NOT YET CODE REVIEWED
}
