% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network_matrix.R
\name{plot_network_matrix}
\alias{plot_network_matrix}
\title{Plot network matrix}
\usage{
plot_network_matrix(
  mat,
  mat_col_palette = NULL,
  ind_names = NULL,
  ind_cols = NULL,
  zlim = NULL,
  xlab = "",
  ylab = "",
  main = ""
)
}
\arguments{
\item{mat}{the matrix to plot (matrix of dimensions \verb{n_inds x n_inds})}

\item{mat_col_palette}{color palette to use for the matrix plot}

\item{ind_names}{the names for the x and y labels (vector of length \code{n_inds})}

\item{ind_cols}{colors to use for the labels (vector of length \code{n_inds})}

\item{zlim}{a vector of length 2 giving the numerical limits of z (the color scale)}

\item{xlab}{x axis label (rows of matrix)}

\item{ylab}{y axis label (columns of matrix)}
}
\description{
Plot a matrix with color representing the edge weights of a network
}
\author{
Ariana Strandburg-Peshkin (primary author)

NOT YET CODE REVIEWED
}
