% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_y_conditioned_on_x.R
\name{plot_y_conditioned_on_x}
\alias{plot_y_conditioned_on_x}
\title{Plot a statistic of y conditioned on the value of x}
\usage{
plot_y_conditioned_on_x(
  x,
  y,
  bins = NULL,
  stat = "median",
  bin_by_quantile = T,
  n_bins = 10,
  error_bar_range = 0.25,
  xlab = "",
  ylab = "",
  main = "",
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{x}{vector of numbers whose value will be conditioned on}

\item{y}{vector of numbers from which a statistic will be computed for values associated with each \code{x} bin}

\item{bins}{bins to use (will override \code{bin_by_quantile} and \code{n_bins} if specified)}

\item{stat}{which statistic of the distribution of \code{y} to use (can be \verb{'mean' or 'median'})}

\item{bin_by_quantile}{specifies whether to bin based on quantiles of the overall distribution of \code{x} (\code{T} or \code{F})}

\item{n_bins}{number of bins to create if \code{bin_by_quantile = T} (bins will be evenly spaced according to quantiles)}

\item{error_bar_range}{if not \code{NULL}, will produce error bars spanning a given quantile range of the distribution of \code{y} for each bin. should be between \code{0} and \code{0.5} (defaults to \code{0.25}, or interquantile range)}

\item{xlab}{label for x axis}

\item{ylab}{label for y axis}

\item{main}{main plot label}

\item{xlim}{if specified, sets limits of x axis}

\item{ylim}{if specified, sets limits of y axis}
}
\value{
Creates a plot, and also outputs a list with \code{bins} (bins used), \code{mids} (the midpoints
of the \code{x} bins), \code{y_stats} (the mean or median of \code{y} for each bin), \code{y_uppers}
and \code{y_lowers} (the upper and lower qauntiles of \code{y} for each bin as specified
by \code{error_bar_range}) and \code{ns} (the number of data points in each bin)
}
\description{
For paired measurements \verb{(x,y)}, plot a specified statistic \code{stat} (e.g. the mean) of
\code{y} for values of \code{x} within bins specified by the vector \code{bins}
}
\details{
For example, if \code{stat = 'mean'}, the function will plot \code{vals[i]} vs. \code{bins[i]} where
\code{vals[i]} is defined as \code{mean(y[which(x >= bins[i] & x < bins[i+1])], na.rm=T)}
}
\author{
Ariana Strandburg-Peshkin

NOT YET CODE REVIEWED
}
