% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_matrix_to_df.R
\name{reformat_matrix_to_df}
\alias{reformat_matrix_to_df}
\title{Convert from matrix format to dataframe}
\usage{
reformat_matrix_to_df(
  xs,
  ys,
  timestamps,
  ids = NULL,
  time_indices_selected,
  id_codes_selected = NULL,
  lats = NULL,
  lons = NULL
)
}
\arguments{
\item{xs}{UTM eastings matrix (\code{n_inds} x \code{n_times} matrix where xs\link{i,t} gives the easting of individual i at time step t)}

\item{ys}{UTM northings matrix (\code{n_inds} x \code{n_times} matrix where ys\link{i,t} gives the northing of individual i at time step t)}

\item{timestamps}{vector of timestamps (POSIXct), must have same dimensions as columns of \code{xs} and \code{ys} matrices}

\item{ids}{Dataframe of individuals with individual name stored in a column called \code{id_code}. Only necessary if you want to select a subset of individuals using \code{id_codes_selected}}

\item{time_indices_selected}{Time indices for data to include in the dataframe, to be matched with xs, ys, timestamps}

\item{id_codes_selected}{Identity codes of individuals to include in dataframe. If none provided, all individuals are included}

\item{lats}{matrix of latitude values (\code{n_inds} x \code{n_times})}

\item{lons}{matrix of longitude values (\code{n_inds} x \code{n_times})}
}
\value{
Returns a dataframe of id_codes, xs, ys, and timestamps
}
\description{
Creates a dataframe out of GPS data in "matrix format" suitable for plotting with ggplot
}
\author{
Eli Strauss

NOT YET CODE REVIEWED
}
