% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_movebank_to_matrix.R
\name{reformat_movebank_to_matrix}
\alias{reformat_movebank_to_matrix}
\title{Reformat Movebank to matrix format}
\usage{
reformat_movebank_to_matrix(
  movebank_data,
  output_file_path = NULL,
  data_chunks = NULL,
  seconds_per_time_step = 1,
  start_date = NULL,
  end_date = NULL,
  start_time = NULL,
  end_time = NULL,
  utm_zone = NULL,
  hemisphere = NULL,
  output_utm = T,
  output_latlon = T
)
}
\arguments{
\item{movebank_data}{data frame that must include the columns
\code{'individual_local_identifier'} (individual id of the tracked animals,
\code{'study_local_timestamp'} (local timestamp for each GPS fix, must be POSIXct objects including timezone),
\code{'location_long'} (longitude coordinate), and
\verb{'location_lat} (latitude coordinate)}

\item{output_file_path}{full path to the desired output file}

\item{data_chunks}{data frame specifying starting and ending datetimes (in POSIXct format - with time zone!) for each contiguous chunk of data. overrides other specified start and end times / dates}

\item{seconds_per_time_step}{sampling interval of GPS fixes (in seconds)}

\item{start_date}{a character string specifying the starting date for data sampled in a daily basis, only used if \code{data_chunks = NULL}, format must be \code{'YYYY-MM-DD'}}

\item{end_date}{a character string specifying the end date for data sampled in a daily basis, only used if \code{data_chunks = NULL}, format must be \code{'YYYY-MM-DD'}}

\item{start_time}{a character string specifying the end time for data sampled on a daily basis, only used if \code{data_chunks = NULL}, format must be \code{'HH:MM:SS'}}
}
\description{
Takes in regularly sampled GPS data in Movebank format (data frame with specified columns)
and converts it into the standard matrix form used by the cocomo package.
}
\author{
Ariana Strandburg-Peshkin (primary author)

NOT YET CODE REVIEWED
}
