% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm_to_latlon.R
\name{utm_to_latlon}
\alias{utm_to_latlon}
\title{UTM to latitude / longitude conversion}
\usage{
utm_to_latlon(easts_norths, utm_zone, hemisphere)
}
\arguments{
\item{easts_norths}{N x 2 matrix of eastings (col 1) and northings (col2)}

\item{utm_zone}{numeric or string value of UTM zone}

\item{hemisphere}{northern or southern hemisphere - specify 'N' or 'S' (not case sensitive)}
}
\value{
lons_lats, an N x 2 matrix of longitudes (col 1) and latitudes (col 2)
}
\description{
Converts a matrix of UTM eastings and northings (eastings first column,
northings second column) to a matrix of lons and lats (lons first column,
lats second column)
}
\author{
Ariana Strandburg-Peshkin (primary author)

NOT YET CODE REVIEWED
}
