% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_movement_and_calls_visualization.R
\name{generate_movement_and_calls_visualization}
\alias{generate_movement_and_calls_visualization}
\title{Generate images for movement and calls visualization}
\usage{
generate_movement_and_calls_visualization(
  xs = NULL,
  ys = NULL,
  timestamps = NULL,
  calls = NULL,
  start_time = NULL,
  end_time = NULL,
  time_step = 1,
  output_dir = NULL,
  tail_time = 10,
  call_persist_time = 5,
  colors_inds = NULL,
  colors_calls = NULL,
  pchs_inds = NULL,
  pchs_calls = NULL,
  show_legend_inds = T,
  show_legend_calls = T,
  legend_loc = "topright",
  show_time = T,
  show_scalebar = T,
  scalebar_size = 100,
  scalebar_loc = "bottomleft",
  scalebar_offset = 20,
  ind_names = NULL,
  bg_color = "black",
  ind_point_size = NULL,
  call_point_size = NULL,
  events = NULL
)
}
\arguments{
\item{xs}{UTM eastings matrix (\code{n_inds} x \code{n_times} matrix where xs\link{i,t} gives the easting of individual i at time step t)}

\item{ys}{UTM northings matrix (\code{n_inds} x \code{n_times} matrix where ys\link{i,t} gives the northing of individual i at time step t)}

\item{timestamps}{vector of timestamps (POSIXct), must have same dimensions as columns of \code{xs} and \code{ys} matrices}

\item{calls}{data frame where first column (\code{'ind_idx'}) specifies the index of the individual that gave the call, second column (\code{'time_idx'}) specifies the time index at which the call was given, and third column (\code{'call_type'}) specifies the type of call (character string)}

\item{start_time}{time index at which to start the video}

\item{end_time}{time index at which to end the video}

\item{time_step}{time step to use (an image every time_step timesteps will be produced)}

\item{output_dir}{directory in which to store the folder of outputted images}

\item{tail_time}{number of previous time steps to plot as a "tail" which trails the point showing the current location}

\item{call_persist_time}{number of previous time steps to still show the calls (they will shrink linearly over time in size)}

\item{colors_inds}{vector of colors to use for each individual (length \code{n_inds})}

\item{colors_calls}{vector of colors to use for each call type (alphabetical order by call_type)}

\item{pchs_inds}{vector of plotting symbols for individuals}

\item{pchs_calls}{vector of plotting symbols for calls}

\item{show_legend_inds}{whether to plot a legend showing the names of the individuals (T or F)}

\item{show_legend_calls}{whether to plot a legend showing the names of the call types (T or F)}

\item{legend_loc}{location of the legend, either \verb{topleft','topright','bottomleft' or 'bottomright'}}

\item{show_time}{whether to show the timestamp or not (\code{T} or \code{F})}

\item{show_scalebar}{whether to show a scalebar or not (\code{T} or \code{F})}

\item{scalebar_size}{number of meters for the scalebar}

\item{scalebar_loc}{location of the scale bar, either \verb{topleft','topright','bottomleft' or 'bottomright'}}

\item{scalebar_offset}{scalebar offset from the edge (fraction of entire width)}

\item{ind_names}{vector of names of the individuals}

\item{bg_color}{background color}

\item{ind_point_size}{size of the individual points}

\item{call_point_size}{size of the points for calls}

\item{events}{data frame with columns \code{event_id}, \code{inds_involved}, \code{start_time_idx},\code{end_time_idx},\code{highlight_time_idx},\code{inds_highlighted},\code{initiator}}
}
\description{
Generate a bunch of images (PNGs) showing trajectory data over a given time period, optionally on top of a Google Earth map,
and optionally with calls plotted on top.
These are saved as 1.png, 2.png, 3.png etc. to a folder, and you can then use the following ffmpeg command (or similar)
to merge them into a video:
ffmpeg -framerate 2 -i \%d.png -r 8 -c:v libx264 -pix_fmt yuvj420p output_file_name.mp4
Data must be entered either as \code{xs} and \code{ys} matrices (UTM) if \code{plot_on_map = F} or as \code{lats} and \code{lons} matrices (latitude / longitude) if \code{plot_on_map = T}
}
\author{
Ariana Strandburg-Peshkin (primary author)

NOT YET CODE REVIEWED
}
