% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pulls_and_anchors.R
\name{get_pulls_and_anchors}
\alias{get_pulls_and_anchors}
\title{Get pulls and anchors}
\usage{
get_pulls_and_anchors(xa, xb, ya, yb, a, b, noise_thresh = 5, plot_results = F)
}
\arguments{
\item{xa:}{x coordinates for individual a (vector of length \code{n_times})- must be continuous data (no NAs)}

\item{xb:}{x coordinates for individual b (vector of length \code{n_times})- must be continuous data (no NAs)}

\item{ya:}{x coordinates for individual a (vector of length \code{n_times})- must be continuous data (no NAs)}

\item{yb:}{x coordinates for individual b (vector of length \code{n_times})- must be continuous data (no NAs)}

\item{a:}{index of the first individual}

\item{b:}{index of the second individual}

\item{noise_thresh:}{noise threshold (defaults to 5 m)}

\item{plot_results:}{whether to plot results or not}
}
\value{
Returns a data frame  containing dyadic interactions between a and b.
Contains columns: \code{t1}, \code{t2}, \code{t3}, \code{leader}, \code{follower}, \code{event_type}, \code{disparity}, \code{strength},\code{disparity_additive}, and \code{strength_additive}.
The \code{disparity} and \code{strength} are as defined in Strandburg-Peshkin et al. 2015, whereas
\code{disparity_additive} and \code{strength_additive} are alternative formulations of these metrics that add the components together instead of multiplying them.
}
\description{
Get successful ('pull') and failed ('anchor') dyadic interactions between individuals a and b.
}
\author{
Ariana Strandburg-Peshkin (primary author)

NOT YET CODE REVIEWED
}
