% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_together_sticky.R
\name{get_together_sticky}
\alias{get_together_sticky}
\title{Get whether two individuals are together, using a "sticky" definition}
\usage{
get_together_sticky(together_inner, together_outer)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
