% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlon_to_utm.R
\name{latlon_to_utm}
\alias{latlon_to_utm}
\title{Latitude / longitude to UTM conversion}
\usage{
latlon_to_utm(lons_lats, utm_zone, southern_hemisphere)
}
\arguments{
\item{lons_lats}{N x 2 matrix of longitudes (col 1) and latitudes (col2)}
}
\value{
easts_norths, an N x 2 matrix of eastings (col 1) and northings (col 2)
}
\description{
Converts a matrix of lons and lats (lons first column, lats second column) to UTM
eastings and northings (eastings first column, northings second column)
}
\author{
Ariana Strandburg-Peshkin (primary author)

NOT YET CODE REVIEWED
}
