% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pull_or_anchor_event.R
\name{plot_pull_or_anchor_event}
\alias{plot_pull_or_anchor_event}
\title{Plot pull or anchor event}
\usage{
plot_pull_or_anchor_event(xs, ys, events, event_idx, ids = NULL)
}
\arguments{
\item{xs}{\verb{n_inds x n_times} matrix of x positions}

\item{ys}{\verb{n_inds x n_times} matrix of y positions}

\item{events}{events table output from \code{get_pulls_and_anchors}}

\item{event_idx}{row in the table to plot}

\item{ids}{ids table output from \code{reformat_movebank_to_matrix}}
}
\value{
Returns a plot of the event with the initiator in blue and the responder in red
}
\description{
Plots a pull or anchor event, with the initiator colored blue and the responder colored red.
Panels show the positions at \code{t1},\code{t2}, and \code{t3} respectively.
}
\author{
Ariana Strandburg-Peshkin (primary author)

NOT YET CODE REVIEWED
}
