% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_movebank_to_matrix.R
\name{reformat_movebank_to_matrix}
\alias{reformat_movebank_to_matrix}
\title{Reformat Movebank to matrix format}
\usage{
reformat_movebank_to_matrix(
  movebank_data,
  output_file_path = NULL,
  ids = NULL,
  data_chunks = NULL,
  seconds_per_time_step = 1,
  start_date = NULL,
  end_date = NULL,
  start_time = NULL,
  end_time = NULL,
  utm_zone = NULL,
  hemisphere = NULL,
  output_utm = T,
  output_latlon = T,
  use_UTC = T,
  local_timezone = NULL
)
}
\arguments{
\item{movebank_data}{data frame that must include the columns
\verb{'timestamp} (timestamp in UTC, must be character string of format YYYY-MM-DD HH:MM:SS.SSS),
\code{'individual.local.identifier'} (individual id of the tracked animals),
\code{'location.long'} (longitude coordinate), and
\verb{'location.lat} (latitude coordinate).
Can optionally also include the column \code{'study.local.timestamp'} (local timestamp for each GPS fix, must be character string of format YYYY-MM-DD HH:MM:SS.SSS).}

\item{output_file_path}{full path to the desired output file (must be a .RData file)}

\item{ids}{optional ids data frame which if specified will determine the order of rows in matrices (\code{code} column in ids needs to match to \code{individual.local.identifier} column in \code{movebank_data})}

\item{data_chunks}{data frame with columns \code{start} and \code{end} specifying starting and ending datetimes (in POSIXct format - with time zone!) for each contiguous chunk of data. overrides other specified start and end times / dates.}

\item{seconds_per_time_step}{sampling interval of GPS fixes (in seconds)}

\item{start_date}{a character string specifying the starting date for data sampled in a daily basis, only used if \code{data_chunks = NULL}, format must be \code{'YYYY-MM-DD'}}

\item{end_date}{a character string specifying the end date for data sampled in a daily basis, only used if \code{data_chunks = NULL}, format must be \code{'YYYY-MM-DD'}}

\item{start_time}{a character string specifying the start time for data sampled on a daily basis, only used if \code{data_chunks = NULL}, format must be \code{'HH:MM:SS'}}

\item{end_time}{a character string specifying the end time for data sampled on a daily basis, only used if \code{data_chunks = NULL}, format must be \code{'HH:MM:SS'}}

\item{utm_zone}{numeric UTM zone (only needed if \code{output_utm = T} and if \code{movebank_data} does not have \code{utm_easting} and \code{utm_northing} columns)}

\item{hemisphere}{hemisphere (\code{'north'} or \code{'south'}) for UTM calculations (only needed if \code{output_utm = T} and if \code{movebank_data} does not have \code{utm_easting} and \code{utm_northing} columns)}

\item{output_utm}{whether to output \code{xs} and \code{ys} matrices (\code{T} or \code{F})}

\item{output_latlon}{whether to output \code{lats} and \code{lons} matrices (\code{T} or \code{F})}

\item{use_UTC}{if T (default) use UTC time rather than local time}

\item{local_timezone}{specify local timezone string, if not using UTC (use_UTC = F)}
}
\value{
Saves a file to the location \code{output_file_path} containing the objects: \code{timestamps} (vector of timestamps in POSIXct format, of length \code{n_times}),
\code{ids} (data frame containing either the original \code{ids} data or a single \code{code} column with the id from \code{individual.local.identifier} column in \code{movebank_data}), \code{xs} (\code{n_inds} x \code{n_times} matrix of UTM eastings for all individuals at each time point),
\code{ys} (\code{n_inds} x \code{n_times} matrix of UTM northings for all individuals at each time point), \code{lons} (\code{n_inds} x \code{n_times} matrix of longitudes for all individuals at each time point), and
\code{lats} (\code{n_inds} x \code{n_times} matrix of longitudes for all individuals at each time point). \code{xs} and \code{ys} are only saved if \code{output_utm = T}. \code{lons} and \code{lats} are only saved if \code{output_lonlat = T}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Takes in regularly sampled GPS data in Movebank format (i.e. data frame with specified columns - see below)
and converts it into the standard matrix form used by the \code{cocomo} package.
}
\author{
Ariana Strandburg-Peshkin (primary author)

NOT YET CODE REVIEWED
}
