% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synch_audio_file_labels_to_UTC.R
\name{synch_audio_file_labels_to_UTC}
\alias{synch_audio_file_labels_to_UTC}
\title{Synch labels within audio file to UTC time}
\usage{
synch_audio_file_labels_to_UTC(path_to_label_file, path_to_synch_file)
}
\arguments{
\item{path_to_synch_file}{}
}
\description{
Reads in a label file (in Audition format) and a synch file, get synch points,
and synchs all labels in the file to UTC. Outputs a table with an additional
column specifying timestamp_UTC
}
